requirements.txt
dash==2.6.*
Flask==2.1.3
psycopg2-binary==2.9.3
pandas==1.2.4
numpy==1.19.2
matplotlib==3.4.1
plotly==5.5.0
gunicorn
